+++ dataobjects/QuadTree.java
@@ -18,14 +18,14 @@ public class QuadTree<T extends KrakEdge> implements Serializable {
 	 */
 	private QuadTreeNode<T> root;
 
-	public QuadTree(Rectangle2D.Double bounds, Set<T> content){
+	public QuadTree(Rectangle2D.Float bounds, Set<T> content){
 		Stopwatch timer = new Stopwatch("Creating QuadTree");
 		root = new QuadTreeNode<T>(bounds,content);
 		timer.printTime();
 	}
 
 
-	public Set<T> query(Rectangle2D.Double qarea){
+	public Set<T> query(Rectangle2D.Float qarea){
 		return root.query(qarea);
 	}
 }
