+++ utils/PointMethods.java
@@ -15,14 +15,14 @@ public class PointMethods {
 	 * @param v The view the point originated in.
 	 * @return The UTM point.
 	 */
-	public static Point2D.Double pixelToUTM(Point e, Model model, View view){
+	public static Point2D.Float pixelToUTM(Point e, Model model, View view){
 		//Inverts the y-value of the point, so that it is converted from the pixel coordinate system to the
 		//UTM coordinate system
 		e.y = view.getCanvasHeight() - e.y;
 		// Convert pixel to meters
 		float x_m = (float) (model.getBounds().x + (e.getX() / (float) view.getCanvasWidth()) * model.getBounds().width);
 		float y_m = (float) (model.getBounds().y + (e.getY() / (float) view.getCanvasHeight()) * model.getBounds().height);
-		return new Point2D.Double(x_m, y_m);	
+		return new Point2D.Float(x_m, y_m);	
 	}
  
 	/**
@@ -45,7 +45,7 @@ public class PointMethods {
 		}
 	}
 	
-	public static Point UTMToPixel(Point2D.Double e, Model model, View view){
+	public static Point UTMToPixel(Point2D.Float e, Model model, View view){
 		int x_m = (int) (((e.getX() - model.getBounds().x) / model.getBounds().width) * view.getCanvasWidth());
 		int y_m = (int) (((e.getY() - model.getBounds().y) / model.getBounds().height) * view.getCanvasHeight());
